+++ src/tech/lcw.h
@@ -40,7 +40,7 @@
 #ifndef CNC_RED_ALERT_TECH_LCW_H_
 #define CNC_RED_ALERT_TECH_LCW_H_
 
-int LCW_Uncomp(const void* source, void* dest, unsigned long length = 0);
+int LCW_Uncomp(const void* source, void* dest, int length = 0);
 
 extern "C" {
 int __cdecl LCW_Comp(const void* source, void* dest, int length);
